rus/QuizBot/requirements.txt
aiogram==3.22.0
python-dotenv==1.1.1
aiofiles==24.1.0